+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/servers/DockerComposeServer.java
@@ -7,25 +7,26 @@ package de.gematik.test.tiger.testenvmgr.servers;
 import de.gematik.rbellogger.util.RbelAnsiColors;
 import de.gematik.test.tiger.common.Ansi;
 import de.gematik.test.tiger.common.config.ServerType;
+import de.gematik.test.tiger.common.config.TigerConfigurationException;
 import de.gematik.test.tiger.common.data.config.CfgDockerOptions;
-import de.gematik.test.tiger.common.data.config.tigerProxy.TigerRoute;
 import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
 import de.gematik.test.tiger.testenvmgr.config.CfgServer;
 import de.gematik.test.tiger.testenvmgr.env.TigerServerStatusUpdate;
-import lombok.Builder;
-import lombok.extern.slf4j.Slf4j;
-
 import java.util.Collections;
 import java.util.List;
-
-import static de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr.HTTP;
+import lombok.Builder;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 
 @Slf4j
 public class DockerComposeServer extends TigerServer {
 
     @Builder
     DockerComposeServer(String serverId, CfgServer configuration, TigerTestEnvMgr tigerTestEnvMgr) {
-        super(determineHostname(configuration, serverId), serverId, tigerTestEnvMgr, configuration);
+        super("", serverId, tigerTestEnvMgr, configuration);
+        if (!StringUtils.isBlank(configuration.getHostname())) {
+            throw new TigerConfigurationException("Hostname property is not supported for docker compose nodes!");
+        }
     }
 
     @Override
@@ -38,16 +39,6 @@ public class DockerComposeServer extends TigerServer {
             getHostname(), getDockerSource());
         statusMessage("Starting docker compose");
         getTigerTestEnvMgr().getDockerManager().startComposition(this);
-
-        // add routes needed for each server to local docker proxy
-        // ATTENTION only one route per server!
-        if (getConfiguration().getDockerOptions().getPorts() != null
-            && !getConfiguration().getDockerOptions().getPorts().isEmpty()) {
-            addRoute(TigerRoute.builder()
-                .from(HTTP + getHostname())
-                .to(HTTP + "localhost:" + getConfiguration().getDockerOptions().getPorts().values().iterator().next())
-                .build());
-        }
         log.info(Ansi.colorize("Docker compose Startup for {} : {} OK", RbelAnsiColors.GREEN_BOLD),
             getHostname(), getDockerSource());
         statusMessage("Docker compose started");