+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/TestEnvManagerPositive.java
@@ -14,7 +14,10 @@ import de.gematik.test.tiger.testenvmgr.servers.TigerServer;
 import de.gematik.test.tiger.testenvmgr.util.TigerTestEnvException;
 import java.io.File;
 import java.io.IOException;
+import java.net.URI;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import kong.unirest.Unirest;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
@@ -36,6 +39,7 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
     @ParameterizedTest
     @ValueSource(strings = {"testDocker", "testTigerProxy", "testExternalJar", "testExternalUrl"})
     public void testCheckCfgPropertiesMinimumConfigPasses_OK(String cfgFileName) {
+        log.info("Starting testCheckCfgPropertiesMinimumConfigPasses_OK for {}", cfgFileName);
         TigerGlobalConfiguration.initializeWithCliProperties(Map.of("TIGER_TESTENV_CFGFILE",
             "src/test/resources/de/gematik/test/tiger/testenvmgr/" + cfgFileName + ".yaml"));
 
@@ -46,8 +50,9 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"testDockerMVP", "testTigerProxy", "testExternalJarMVP", "testExternalUrl"})
+    @ValueSource(strings = {"testComposeMVP", "testDockerMVP", "testTigerProxy", "testExternalJarMVP", "testExternalUrl"})
     public void testSetUpEnvironmentNShutDownMinimumConfigPasses_OK(String cfgFileName) throws IOException {
+        log.info("Starting testSetUpEnvironmentNShutDownMinimumConfigPasses_OK for {}", cfgFileName);
         FileUtils.deleteDirectory(new File("WinstoneHTTPServer"));
         createTestEnvMgrSafelyAndExecute(envMgr -> {
             envMgr.getConfiguration().getServers().get(cfgFileName);
@@ -72,8 +77,7 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         TigerGlobalConfiguration.initializeWithCliProperties(Map.of("TIGER_TESTENV_CFGFILE",
             "src/test/resources/de/gematik/test/tiger/testenvmgr/testComposeWithHostname.yaml"));
 
-        createTestEnvMgrSafelyAndExecute(envMgr ->
-            assertThatThrownBy(envMgr::setUpEnvironment).isInstanceOf(TigerConfigurationException.class));
+        assertThatThrownBy(TigerTestEnvMgr::new).isInstanceOf(TigerConfigurationException.class);
     }
 
     // -----------------------------------------------------------------------------------------------------------------
@@ -84,10 +88,35 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
     @Test
     public void testCreateDockerNonExistingVersion() {
         TigerGlobalConfiguration.initializeWithCliProperties(Map.of("TIGER_TESTENV_CFGFILE",
-                "src/test/resources/de/gematik/test/tiger/testenvmgr/testDockerMVP.yaml",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDockerMVP.yaml",
             "tiger.servers.testDockerMVP.version", "200.200.200-2000"));
         createTestEnvMgrSafelyAndExecute(envMgr ->
             assertThatThrownBy(envMgr::setUpEnvironment).isInstanceOf(TigerTestEnvException.class));
+    }
+
+    @Test
+    public void testCreateDockerComposeAndCheckPortIsAvailable() throws IOException {
+        createTestEnvMgrSafelyAndExecute(envMgr -> {
+            envMgr.setUpEnvironment();
+            String host = System.getenv("DOCKER_HOST");
+            if (host == null) {
+                host = "localhost";
+            } else {
+                host = new URI(host).getHost();
+            }
+            log.info("Web server expected to serve at {}", TigerGlobalConfiguration.resolvePlaceholders("http://" + host + ":${free.port.1}"));
+            try {
+                log.info("Web server responds with: " +
+                    Unirest.spawnInstance()
+                        .get(TigerGlobalConfiguration.resolvePlaceholders("http://" + host + ":${free.port.1}"))
+                        .asString().getBody());
+            } catch (Exception e) {
+                log.error("Unable to retrieve document from docker compose webserver...", e);
+            }
+            assertThat(Unirest.spawnInstance().get(
+                    TigerGlobalConfiguration.resolvePlaceholders("http://" + host +":${free.port.1}"))
+                .asString().getStatus()).isEqualTo(200);
+        }, "src/test/resources/de/gematik/test/tiger/testenvmgr/testComposeMVP.yaml");
     }
 
     // -----------------------------------------------------------------------------------------------------------------
@@ -172,7 +201,7 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
             } finally {
                 FileUtils.forceDeleteOnExit(folder);
             }
-        },  "src/test/resources/de/gematik/test/tiger/testenvmgr/testExternalJarMVP.yaml");
+        }, "src/test/resources/de/gematik/test/tiger/testenvmgr/testExternalJarMVP.yaml");
     }
 
     @Test
@@ -226,7 +255,6 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
             .hasMessageStartingWith("Local jar ").hasMessageEndingWith("miniJarWHICHDOESNOTEXIST.jar not found!");
     }
 
-
     // -----------------------------------------------------------------------------------------------------------------
     //
     // local tiger proxy details
@@ -246,8 +274,10 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         "        - \"--webroot=.\"\n",
         skipEnvironmentSetup = true)
     public void startLocalTigerProxyAndCheckPropertiesSet(TigerTestEnvMgr envMgr) {
-        assertThat(TigerGlobalConfiguration.readIntegerOptional(TigerTestEnvMgr.CFG_PROP_NAME_LOCAL_PROXY_ADMIN_PORT).get()).isBetween(0, 655536);
-        assertThat(TigerGlobalConfiguration.readIntegerOptional(TigerTestEnvMgr.CFG_PROP_NAME_LOCAL_PROXY_PROXY_PORT).get()).isBetween(0, 655536);
+        assertThat(TigerGlobalConfiguration.readIntegerOptional(TigerTestEnvMgr.CFG_PROP_NAME_LOCAL_PROXY_ADMIN_PORT).get()).isBetween(0,
+            655536);
+        assertThat(TigerGlobalConfiguration.readIntegerOptional(TigerTestEnvMgr.CFG_PROP_NAME_LOCAL_PROXY_PROXY_PORT).get()).isBetween(0,
+            655536);
     }
 
     @Test